Inline AbstractManagedChannelImplBuilder: convert methods calling super()
